added comment to the function

+++ 0x00-python_variable_annotations/2-floor.py
@@ -6,6 +6,6 @@ import math
 """
 
 
-def floor(n: float) -> int:
+def floor(n: float) -> float:
     """Returns the floor of the float n"""
     return math.floor(n)
